+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# [18-02-2023]
-
+# [18-02-2023] 0.1.1
+## Added
+- Setup rocket runtime
+- Create first two routes for users signup and login (wihtout funcionality)
+# [18-02-2023] 0.1.0
 ## Added
 - Created project